Drafted Manager UI
The display has 2 sections - the left hand column view is a user action (select date range and a report); the right hand table view is the pull from the DB as the result of the user action. 
Note: only the Sales Report table view has been built. Need to build 4 others, but need to understand the trigger mechanism and how to display / hide based on user action.